+++ site/src/modules/dashboard/DeploymentBanner/DeploymentBanner.tsx
@@ -2,7 +2,7 @@ import type { FC } from "react";
 import { useQuery } from "react-query";
 import { health } from "api/queries/debug";
 import { deploymentStats } from "api/queries/deployment";
-import { useAuthenticated } from "contexts/auth/useAuth";
+import { useAuthenticated } from "contexts/auth/RequireAuth";
 import { DeploymentBannerView } from "./DeploymentBannerView";
 
 export const DeploymentBanner: FC = () => {
