+++ site/src/pages/DeploySettingsPage/DeploySettingsLayout.tsx
@@ -7,7 +7,7 @@ import { Loader } from "components/Loader/Loader";
 import { Margins } from "components/Margins/Margins";
 import { Stack } from "components/Stack/Stack";
 import { RequirePermission } from "contexts/auth/RequirePermission";
-import { useAuthenticated } from "contexts/auth/useAuth";
+import { useAuthenticated } from "contexts/auth/RequireAuth";
 import { Sidebar } from "./Sidebar";
 
 type DeploySettingsContextValue = {
