+++ site/src/pages/TemplatesPage/TemplatesPage.tsx
@@ -2,13 +2,12 @@ import type { FC } from "react";
 import { Helmet } from "react-helmet-async";
 import { useQuery } from "react-query";
 import { templateExamples, templates } from "api/queries/templates";
-import { useAuthenticated } from "contexts/auth/useAuth";
+import { useAuthenticated } from "contexts/auth/RequireAuth";
 import { pageTitle } from "utils/page";
 import { TemplatesPageView } from "./TemplatesPageView";
 
 export const TemplatesPage: FC = () => {
-  const { orgId } = useAuthenticated();
-  const { permissions } = useAuthenticated();
+  const { orgId, permissions } = useAuthenticated();
   const templatesQuery = useQuery(templates(orgId));
   const examplesQuery = useQuery({
     ...templateExamples(orgId),
