+++ site/src/pages/UserSettingsPage/AppearancePage/AppearancePage.tsx
@@ -3,7 +3,7 @@ import type { FC } from "react";
 import { useMutation, useQueryClient } from "react-query";
 import { updateAppearanceSettings } from "api/queries/users";
 import { Stack } from "components/Stack/Stack";
-import { useAuthenticated } from "contexts/auth/useAuth";
+import { useAuthenticated } from "contexts/auth/RequireAuth";
 import { Section } from "../Section";
 import { AppearanceForm } from "./AppearanceForm";
